pyproject.toml
[project]
name = "gen-slide-chaper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "kanta", email = "101532773+kantas-spike@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = []

[project.scripts]
gen-slide-chaper = "gen_slide_chaper:main"

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "fake-bpy-module>=20251003",
    "pytest>=8.4.2",
    "ruff>=0.14.0",
]
